BeerBuddy-Python/BeerBuddy/requirements.txt
apns2==0.4.1
asgiref==3.3.1
asn1crypto==0.24.0
backcall==0.1.0
certifi==2018.10.15
cffi==1.14.3
chardet==3.0.4
clx-sdk-xms==1.0.0
cryptography==2.3.1
decorator==4.3.0
dj-database-url==0.5.0
Django==3.1.4
django-appconf==1.0.2
django-ckeditor==5.6.1
django-crontab==0.7.1
django-extensions==2.1.3
django-filter==2.0.0
django-imagekit==4.0.2
django-js-asset==1.2.2
django-mysql==3.9.0
django-utils-six==2.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
docutils==0.14
geographiclib==1.49
geopy==1.17.0
h2==2.6.2
hpack==3.0.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.7
ipdb==0.11
ipython==7.1.1
ipython-genutils==0.2.0
iso8601==0.1.12
jedi==0.13.1
jmespath==0.9.3
mysqlclient==1.3.13
orderedset==2.0.3
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
pilkit==2.0
Pillow==5.3.0
prompt-toolkit==2.0.6
psycopg2==2.8.6
psycopg2-binary==2.8.3
ptyprocess==0.6.0
pycparser==2.19
pyfcm==1.4.5
Pygments==2.2.0
PyJWT==1.6.4
PyMySQL
python-dateutil==2.7.3
python-decouple==3.1
pytz==2020.4
requests==2.20.0
requests-toolbelt==0.8.0
s3transfer==0.1.13
six==1.15.0  
sqlparse==0.4.1
traitlets==4.3.2
urllib3==1.24
wcwidth==0.1.7
Werkzeug==0.14.1
gunicorn==19.6.0
boto3
django-storages==1.12
whitenoise



